feat: 관련된 상장 데이터 없을때 처리
