+++ src/pages/AwardArchivePage/components/AwardTable.tsx
@@ -1,7 +1,5 @@
-import { EmptyBoxImg } from '@/assets';
-import { Flex, Heading, Table } from '@/components';
+import { Table } from '@/components';
 import { THeader } from '@/types/table';
-import { useTheme } from '@mui/material';
 
 import { AwardResponse } from '../types/award';
 
@@ -10,7 +8,6 @@ interface Props {
 }
 
 export const AwardTable = ({ awardList }: Props) => {
-  if (!awardList.length) return <EmptyTable />;
   return <Table headItems={headerItems} bodyItems={awardList} />;
 };
 
@@ -21,24 +18,3 @@ const headerItems: THeader[] = [
   { text: '주관', value: 'organization', width: '200px' },
   { text: '발급일자', value: 'awardDate', width: '200px' },
 ];
-
-const EmptyTable = () => {
-  const theme = useTheme();
-  return (
-    <Flex.Column
-      width="100%"
-      height="400px"
-      gap="1rem"
-      justify="center"
-      align="center"
-    >
-      <EmptyBoxImg />
-      <Heading
-        as="h2"
-        style={{ fontSize: '2rem', color: theme.palette.grey300 }}
-      >
-        등록된 상장이 없어요
-      </Heading>
-    </Flex.Column>
-  );
-};
