+++ src/pages/AwardArchivePage/components/AwardTableListSection.tsx
@@ -1,11 +1,13 @@
 import { Flex, Title } from '@/components';
 
+import { EmptyTable } from '../components/EmptyTable';
 import { useGroupedAwardList } from '../hooks/useGroupedAwardList';
 import { AwardTable } from './AwardTable';
 
 export const AwardTableListSection = () => {
   const { groupedAwardList } = useGroupedAwardList();
 
+  if (!groupedAwardList.length) return <EmptyTable />;
   return (
     <>
       {groupedAwardList.map(group => (
